Copy over nice errors from previous work
