+++ tct/src/tree.rs
@@ -15,6 +15,9 @@ use crate::Witness;
 pub(crate) mod epoch;
 pub(crate) use epoch::block;
 
+pub mod verify;
+use verify::*;
+
 /// A sparse merkle tree witnessing up to 65,536 epochs of up to 65,536 blocks of up to 65,536
 /// [`Commitment`]s.
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
@@ -643,54 +646,4 @@ impl Tree {
             Err(InvalidWitnesses { root, errors })
         }
     }
-}
-
-/// The index for the tree contained at least one error.
-pub struct IndexMalformed {
-    /// The errors found in the index.
-    pub errors: Vec<IndexError>,
-}
-
-pub enum IndexError {
-    /// The index is missing a position.
-    UnindexedWitness {
-        /// The position expected to be present in the index.
-        position: Position,
-        /// The hash found at that position.
-        found_hash: Hash,
-    },
-    /// A commitment in the index doesn't match the hash in the tree at that position.
-    HashMismatch {
-        /// The commitment which should have the found hash.
-        commitment: Commitment,
-        /// The position that commitment maps to in the index.
-        position: Position,
-        /// The expected hash value of that commitment.
-        expected_hash: Hash,
-        /// The actual hash found in the tree structure at the position in the index for that commitment.
-        found_hash: Hash,
-    },
-}
-
-/// At least one proof generated by the tree failed to verify against the root.
-pub struct InvalidWitnesses {
-    /// The root of the tree at which the errors were found.
-    pub root: Root,
-    /// The errors found.
-    pub errors: Vec<WitnessError>,
-}
-
-pub enum WitnessError {
-    /// The index contains a commitment that is not witnessed.
-    UnwitnessedCommitment {
-        /// The commitment that was not present in the tree.
-        commitment: Commitment,
-        /// The position at which it was supposed to appear.
-        position: Position,
-    },
-    /// The proof produced by the tree does not verify against the root.
-    InvalidProof {
-        /// The proof which failed to verify.
-        proof: Box<Proof>,
-    },
 }
