+++ tct/src/tree/verify.rs
@@ -1,0 +1,102 @@
+//! Errors resulting from internal verification methods on [`Tree`]s.
+
+use super::*;
+
+/// The index for the tree contained at least one error.
+#[derive(Clone, Debug, Error)]
+#[error("malformed index:{}", display_errors(.errors))]
+pub struct IndexMalformed {
+    /// The errors found in the index.
+    pub errors: Vec<IndexError>,
+}
+
+fn display_errors(errors: impl IntoIterator<Item = impl Display>) -> String {
+    let mut output = String::new();
+    for error in errors.into_iter() {
+        output.push_str(&format!("\n  {}", error));
+    }
+    output
+}
+
+/// An error occurred when verifying the tree's index.
+#[derive(Clone, Debug, Error)]
+pub enum IndexError {
+    /// The index is missing a position.
+    #[error("unindexed position `{position:?}` with hash {found_hash:?}")]
+    UnindexedWitness {
+        /// The position expected to be present in the index.
+        position: Position,
+        /// The hash found at that position.
+        found_hash: Hash,
+    },
+    /// A commitment in the index doesn't match the hash in the tree at that position.
+    #[error("mismatched hash for commitment {commitment:?} at position `{position:?}`: found {found_hash:?}, expected {expected_hash:?}")]
+    HashMismatch {
+        /// The commitment which should have the found hash.
+        commitment: Commitment,
+        /// The position that commitment maps to in the index.
+        position: Position,
+        /// The expected hash value of that commitment.
+        expected_hash: Hash,
+        /// The actual hash found in the tree structure at the position in the index for that commitment.
+        found_hash: Hash,
+    },
+}
+
+/// At least one proof generated by the tree failed to verify against the root.
+#[derive(Clone, Debug, Error)]
+#[error(
+    "invalid witnesses produced by tree for root {root:?}:{}",
+    display_errors(errors)
+)]
+pub struct InvalidWitnesses {
+    /// The root of the tree at which the errors were found.
+    pub root: Root,
+    /// The errors found.
+    pub errors: Vec<WitnessError>,
+}
+
+/// An error occurred when verifying the tree's contained witnesses.
+#[derive(Clone, Debug, Error)]
+pub enum WitnessError {
+    /// The index contains a commitment that is not witnessed.
+    #[error("unwitnessed commitment {commitment:?} at position `{position:?}`")]
+    UnwitnessedCommitment {
+        /// The commitment that was not present in the tree.
+        commitment: Commitment,
+        /// The position at which it was supposed to appear.
+        position: Position,
+    },
+    /// The proof produced by the tree does not verify against the root.
+    #[error("invalid proof for commitment {:?} at position `{:?}`", .proof.commitment(), .proof.position())]
+    InvalidProof {
+        /// The proof which failed to verify.
+        proof: Box<Proof>,
+    },
+}
+
+/// The tree contained at least one internal cached hash that was incorrect.
+#[derive(Clone, Debug, Error)]
+#[error("invalid cached hashes:{}", display_errors(.errors))]
+pub struct InvalidCachedHashes {
+    /// The errors found in the tree.
+    pub errors: Vec<InvalidCachedHash>,
+}
+
+/// An mismatch between a cached hash and the hash it ought to have been.
+#[derive(Clone, Debug, Error)]
+#[error("cache for `{place}::{kind}` at height {height}, index {index} is incorrect: found {cached:?}, expected {recomputed:?}")]
+pub struct InvalidCachedHash {
+    /// The place of the node with the error.
+    pub place: Place,
+    /// The kind of the node with the error.
+    pub kind: Kind,
+    /// The height of the node with the error.
+    pub height: u8,
+    /// The index of the node with the error.
+    pub index: u64,
+    /// The previous cached hash at that location.
+    pub cached: Hash,
+    /// The recomputed hash that should have been there.
+    pub recomputed: Hash,
+}
